fix: 任务前端不显示 - review_score int 类型拒绝小数分
根因: AgentRecord.review_score 定义为 int，但 reviewer LLM 打分可能返回
小数(如 66.5)，Pydantic 严格验证拒绝带小数的 float(int_from_float)，
TaskState 加载失败 -> get_task 返回 500 -> 前端无法显示任务。

修复: review_score: int | None -> float | None (评分允许小数)
前端 api.ts 已是 number | null，兼容 float，无需改。

验证: 之前 500 的任务(20260717_020109_135219, review_score=66.5)现在 200 正常加载

+++ app/storage/models.py
@@ -82,7 +82,7 @@ class AgentRecord(BaseModel):
     artifact_path: str | None = None      # 相对任务目录的产物路径
     log_path: str | None = None
     retry_count: int = 0
-    review_score: int | None = None       # 审查打分 0-100
+    review_score: float | None = None      # 审查打分 0-100（允许小数，如 66.5）
     review_passed: bool | None = None
     summary: str | None = None            # 一句话摘要，供前端展示
     error: str | None = None
